first

+++ src/controllers/rentals.controllers.js
@@ -1,4 +1,5 @@
 import { connectionDB } from "../database/db.js";
+import dayjs from "dayjs";
 
 export async function create(req,res){
     const{customerId, gameId, daysRented, rentDate, originalPrice, returnDate, delayFee} = res.locals.rentals;
@@ -69,11 +70,41 @@ export async function deleteRental(req,res){
     const {id} = req.params;
     const idRental = await connectionDB.query("SELECT * FROM rentals WHERE id=$1",[id]);
     if(idRental.rowCount === 0){
-        return res.sendStatus(404)
-    } else {
-        await connectionDB.query("DELETE FROM rentals WHERE id=$1",[id]);
-        res.sendStatus(200);
+        return res.sendStatus(404);
+    } 
+    if(!idRental.rows[0].returnDate){
+        return res.sendStatus(404);
     }
 
+    await connectionDB.query("DELETE FROM rentals WHERE id=$1",[id]);
+    console.log(idRental);
+    res.sendStatus(200);
 }
+
+export async function returnRental(req, res) {
+    const { rentDate, daysRented, pricePerDay, id } = res.locals.rental;
+    let delayFee = 0;
+    
+    const returnDate = dayjs();
+    const deadline = dayjs(rentDate).add(daysRented, "day");
+    const isOverdue = dayjs(returnDate).isAfter(deadline);
+  
+    if (isOverdue) {
+      const daysExceeded = (returnDate - deadline) / (1000 * 3600 * 24);
+      delayFee = pricePerDay * Math.ceil(daysExceeded);
+    }
+  
+    try {
+      await connectionDB.query(`
+        UPDATE rentals 
+          SET "returnDate"=$1, "delayFee"=$2
+        WHERE id=$3
+      `, [returnDate.format("YYYY-MM-DD"), delayFee, id]);
+  
+      res.sendStatus(200);
+    } catch (error) {
+      console.log(error);
+      res.sendStatus(500);
+    }
+  }
